refactor: naming refactored, custom errors handling added

+++ src/main/java/com/bookstore/app/entities/product/usecases/DeleteProductUseCase.java
@@ -1,6 +1,7 @@
 package com.bookstore.app.entities.product.usecases;
 
 import com.bookstore.app.entities.product.persistance.IProductsRepository;
+import com.bookstore.app.exceptions.QueryException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +11,8 @@ import java.util.UUID;
 public class DeleteProductUseCase {
     @Autowired
     private IProductsRepository productsRepository;
-    public void handle(UUID id) {
+    public void handle(UUID id) throws QueryException {
+        if (productsRepository.findProductById(id) == null) throw new QueryException("There is no such product");
         productsRepository.deleteProductById(id);
     }
 }
